Move stats into separate module
