+++ src/collector/main.py
@@ -1,9 +1,9 @@
 import logging
 import os
-from datetime import datetime
 
 from src import META_DATA_PATH, ABI_FILEPATH, SNAPSHOT_FILE_PATH
-from src.collector.database.ops import create_tables, get_schain_stats
+from src.collector.core.statistics import aggregate_schain_stats
+from src.collector.database.ops import create_tables
 from src.collector.core.endpoints import get_all_names, is_dkg_passed, get_schain_endpoint
 from src.collector.core.fetchers import Collector, PricesCollector
 from src.utils.helper import daemon, write_json
@@ -11,40 +11,6 @@ from src.utils.logger import init_logger
 from src.utils.meta import create_meta_file, get_meta_file, update_meta_file
 
 logger = logging.getLogger(__name__)
-
-
-def update_dict(target_dict, chain_dict):
-    for key in chain_dict:
-        if type(chain_dict[key]) is dict:
-            if target_dict.get(key) is None:
-                target_dict[key] = {}
-            update_dict(target_dict[key], chain_dict[key])
-        else:
-            target_dict[key] = target_dict.get(key, 0) + chain_dict[key]
-
-
-def aggregate_schain_stats(names):
-    stats = {
-        'schains_number': len(names),
-        'summary': {
-            'total': {},
-            'total_7d': {},
-            'total_30d': {},
-            'group_by_month': {}
-        }
-    }
-    for name in names:
-        logger.info(f'Aggregating stats for {name}')
-        schain_data = get_schain_stats(name)
-        stats.update({
-            name: schain_data
-        })
-        update_dict(stats['summary']['total'], schain_data['total'])
-        update_dict(stats['summary']['total_7d'], schain_data['total_7d'])
-        update_dict(stats['summary']['total_30d'], schain_data['total_30d'])
-        update_dict(stats['summary']['group_by_month'], schain_data['group_by_month'])
-    stats['inserted_at'] = datetime.today().strftime('%Y-%m-%d %H:%M:%S')
-    return stats
 
 
 @daemon(delay=600)
